Add a server component
This is written in Python3, for now.

Previously I was using a generic web server, but in order to use a
real database (like Neo4j) we need a custom server that can handle data
requests on the same port as the website content is served from.

Trying to access the database from the browser directly falls foul of
the "same-origin policy", and is not really well supported by any Neo4j
tooling.