Ensure all relationships are only 1 direction
I.E. remove produced-by, contained-in and required-by.

At the ontology level, it's nice to have these properties, but lower
down they are a pain.

Do you manually specify both? Or do you have a reasoner come in later
and fill in any missing ones? Why bother storing all that duplicated
data in the database when you can query either direction in any decent
graph database?

I thought this would make the JSON-LD content easier to write to be
honest, but (a) it didn't, and (b) you can use @reverse to achieve
the same thing anyway.