import/neo4j: Don't actually try to fetch remote JSON-LD context
I added a link to a public URL of the JSON-LD context in
examples/shared-lib.yaml. I don't want to duplicate the context
everywhere, so it needs to be a link. And there's no way to give a
working link within this app at present.

But, importing from the real Github URL turns out to be broken, probably
because my current version of Python3 doesn't have a working CA
certificates bundle by default. I get this failure:

    ...
    pyld.jsonld.JsonLdError: ('Dereferencing a URL did not result in a valid
    JSON-LD context.',)
    Type: jsonld.ContextUrlError
    Code: loading remote context failed
    Details: {'url':
    'https://raw.githubusercontent.com/ssssam/software-integration-ontology/master/context.jsonld'}
    Cause: ('Could not retrieve a JSON-LD document from the URL.',)
    Type: jsonld.LoadDocumentError
    Code: loading document failed
    Cause: <urlopen error [SSL: CERTIFICATE_VERIFY_FAILED] certificate
    verify failed (_ssl.c:600)>  File
    "/home/sam/.local/lib/python3.4/site-packages/pyld/jsonld.py", line 389,
    in load_document
    ...

So now the importer 'monkey-patches' the URL back to a local file path,
which is a bit rubbish, but OK.